Cleaning up of the git
I have added some functionality in the normal Re-indent file. I hav not
included any GUI yet, but that is just a copy and paste we can do later.

The new part is, to my testings, a pretty reliable string and comment
"mapper". I worked hard on that, the logic was hard to figure out, I had
this rewritten like 3 times :P